Refactor and update
Refactor and update
