+++ automation_editor/__init__.py
@@ -1,0 +1,6 @@
+from automation_editor.automation_editor_ui\
+    .editor_main.main_ui import start_editor
+
+__all__ = [
+    "start_editor"
+]
