--- integration_testing_environment/__init__.py
@@ -1,6 +1,0 @@
-from integration_testing_environment.integration_testing_environment_ui\
-    .editor_main.main_ui import start_editor
-
-__all__ = [
-    "start_editor"
-]
